+++ sdk/core/azure-core/src/test/java/com/azure/core/implementation/serializer/HttpResponseHeaderDecoderTests.java
@@ -38,35 +38,26 @@ public class HttpResponseHeaderDecoderTests {
 
     @Test
     public void nullHeaderTypeReturnsMonoEmpty() {
-        HttpResponseDecodeData decodeData = mock(HttpResponseDecodeData.class);
-        when(decodeData.getHeadersType()).thenReturn(null);
-
-        assertNull(HttpResponseHeaderDecoder.decode(null, null, decodeData));
+        assertNull(HttpResponseHeaderDecoder.decode(null, null, null));
     }
 
     @Test
     public void ioExceptionIsMappedToHttpResponseException() throws IOException {
-        HttpResponseDecodeData decodeData = mock(HttpResponseDecodeData.class);
-        when(decodeData.getHeadersType()).thenReturn(MockHeaders.class);
-
         SerializerAdapter serializer = mock(SerializerAdapter.class);
         when(serializer.deserialize(any(), any())).thenThrow(IOException.class);
 
         HttpResponse response = new MockHttpResponse(null, 200);
 
         assertThrows(HttpResponseException.class,
-            () -> HttpResponseHeaderDecoder.decode(response, serializer, decodeData));
+            () -> HttpResponseHeaderDecoder.decode(response, serializer, MockHeaders.class));
     }
 
     @Test
     public void headersAreDeserializedToType() {
-        HttpResponseDecodeData decodeData = mock(HttpResponseDecodeData.class);
-        when(decodeData.getHeadersType()).thenReturn(MockHeaders.class);
-
         HttpResponse response = new MockHttpResponse(null, 200, new HttpHeaders().set("mock-a", "a"));
 
         Object actual = assertDoesNotThrow(() -> HttpResponseHeaderDecoder.decode(response, new JacksonAdapter(),
-            decodeData));
+            MockHeaders.class));
         assertTrue(actual instanceof MockHeaders);
         MockHeaders mockHeaders = (MockHeaders) actual;
         assertEquals(Collections.singletonMap("a", "a"), mockHeaders.getHeaderCollection());