+++ sdk/resourcemanager/azure-resourcemanager-appplatform/src/main/java/com/azure/resourcemanager/appplatform/models/GitPatternRepository.java
@@ -6,15 +6,12 @@ package com.azure.resourcemanager.appplatform.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
-/** Git repository property payload. */
+/** Git repository property payload for config server. */
 @Fluent
 public final class GitPatternRepository {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(GitPatternRepository.class);
-
     /*
      * Name of the repository
      */
@@ -308,14 +305,16 @@ public final class GitPatternRepository {
      */
     public void validate() {
         if (name() == null) {
-            throw logger
+            throw LOGGER
                 .logExceptionAsError(
                     new IllegalArgumentException("Missing required property name in model GitPatternRepository"));
         }
         if (uri() == null) {
-            throw logger
+            throw LOGGER
                 .logExceptionAsError(
                     new IllegalArgumentException("Missing required property uri in model GitPatternRepository"));
         }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(GitPatternRepository.class);
 }