+++ sdk/resourcemanager/azure-resourcemanager-authorization/src/main/java/com/azure/resourcemanager/authorization/fluent/models/MicrosoftGraphCertificateAuthority.java
@@ -7,7 +7,6 @@ package com.azure.resourcemanager.authorization.fluent.models;
 import com.azure.core.annotation.Fluent;
 import com.azure.core.util.Base64Url;
 import com.azure.core.util.CoreUtils;
-import com.azure.core.util.logging.ClientLogger;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -18,8 +17,6 @@ import java.util.Map;
 /** certificateAuthority. */
 @Fluent
 public final class MicrosoftGraphCertificateAuthority {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(MicrosoftGraphCertificateAuthority.class);
-
     /*
      * Required. The base64 encoded string representing the public certificate.
      */
@@ -72,7 +69,7 @@ public final class MicrosoftGraphCertificateAuthority {
      */
     public byte[] certificate() {
         if (this.certificate == null) {
-            return null;
+            return new byte[0];
         }
         return this.certificate.decodedBytes();
     }