+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ### Other Changes
 
+#### Dependency Updates
+
+- Update `azure-messaging-eventhubs` dependency to `5.11.2`.
+- Update `azure-storage-blob` dependency to `12.16.0`.
+
+## 1.12.1 (2022-04-11)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Update `azure-messaging-eventhubs` dependency to `5.11.2`.
+- Update `azure-storage-blob` dependency to `12.16.0`.
+
 ## 1.12.0 (2022-03-17)
 
 ### Other Changes